Crypto.Secp256k1

Description

Crytpographic functions from Bitcoin’s secp256k1 library.

Messages

data Msg

msg :: ByteString -> Maybe Msg

Import 32-byte ByteString as Msg.

getMsg :: Msg -> ByteString

Get 32-byte message.

Secret Keys

data SecKey

secKey :: ByteString -> Maybe SecKey

Import 32-byte ByteString as SecKey.

getSecKey :: SecKey -> ByteString

Get 32-byte secret key.

derivePubKey :: SecKey -> PubKey

Public Keys

data PubKey

importPubKey :: ByteString -> Maybe PubKey

Import DER-encoded public key.

exportPubKey :: Bool -> PubKey -> ByteString

Encode public key as DER. First argument True for compressed output.

Signatures

data Sig

signMsg :: SecKey -> Msg -> Sig

verifySig :: PubKey -> Sig -> Msg -> Bool

Verify message signature. True means that the signature is correct.

normalizeSig :: Sig -> (Sig, Bool)

Convert signature to a normalized lower-S form. Boolean value True indicates that the signature changed, False indicates that it was already normal.

DER

importSig :: ByteString -> Maybe Sig

Import DER-encoded signature.

laxImportSig :: ByteString -> Maybe Sig

Relaxed DER parsing. Allows certain DER errors and violations.

exportSig :: Sig -> ByteString

Encode signature as strict DER.

Compact

data CompactSig

Constructors

CompactSig

Fields

getCompactSigR :: Word256

getCompactSigS :: Word256

exportCompactSig :: Sig -> CompactSig

importCompactSig :: CompactSig -> Maybe Sig

Addition & Multiplication

data Tweak

tweak :: ByteString -> Maybe Tweak

32-Byte ByteString as Tweak.

getTweak :: Tweak -> ByteString

Get 32-byte tweak.

tweakAddSecKey :: SecKey -> Tweak -> Maybe SecKey

Add tweak to secret key.

tweakMulSecKey :: SecKey -> Tweak -> Maybe SecKey

Multiply secret key by tweak.

tweakAddPubKey :: PubKey -> Tweak -> Maybe PubKey

Add tweak to public key. Tweak is multiplied first by G to obtain a point.

tweakMulPubKey :: PubKey -> Tweak -> Maybe PubKey

Multiply public key by tweak. Tweak is multiplied first by G to obtain a point.

combinePubKeys :: [PubKey] -> Maybe PubKey

Add multiple public keys together.