Crypto.Secp256k1.Internal

Description

The API for this module may change at any time. This is an internal module only exposed for hacking and experimentation.

Documentation

data Ctx

Constructors

Ctx

newtype PubKey64

Constructors

PubKey64

Fields

getPubKey64 :: ByteString

newtype Msg32

Constructors

Msg32

Fields

getMsg32 :: ByteString

newtype Sig64

Constructors

Sig64

Fields

getSig64 :: ByteString

data CompactSig

Constructors

CompactSig

Fields

getCompactSigR :: Word256

getCompactSigS :: Word256

newtype Seed32

Constructors

Seed32

Fields

getSeed32 :: ByteString

newtype SecKey32

Constructors

SecKey32

Fields

getSecKey32 :: ByteString

newtype Tweak32

Constructors

Tweak32

Fields

getTweak32 :: ByteString

newtype Nonce32

Constructors

Nonce32

Fields

getNonce32 :: ByteString

newtype Algo16

Constructors

Algo16

Fields

getAlgo16 :: ByteString

newtype CtxFlags

Constructors

CtxFlags

Fields

getCtxFlags :: CUInt

newtype SerFlags

Constructors

SerFlags

Fields

getSerFlags :: CUInt

newtype Ret

Constructors

Ret

Fields

getRet :: CInt

type NonceFunction a

Arguments

= Ptr Nonce32

-> Ptr Msg32

-> Ptr SecKey32

-> Ptr Algo16

-> Ptr a
extra data

-> CUInt
attempt

-> Ret

Nonce32-generating function

verify :: CtxFlags

sign :: CtxFlags

signVerify :: CtxFlags

compressed :: SerFlags

uncompressed :: SerFlags

useByteString :: ByteString -> ((Ptr CUChar, CSize) -> IO a) -> IO a

packByteString :: (Ptr CUChar, CSize) -> IO ByteString

isSuccess :: Ret -> Bool

ctx :: Ptr Ctx

contextCreate :: CtxFlags -> IO (Ptr Ctx)

contextClone :: Ptr Ctx -> IO (Ptr Ctx)

contextDestroy :: FunPtr (Ptr Ctx -> IO ())

setIllegalCallback

Arguments

:: Ptr Ctx

-> FunPtr (CString -> Ptr a -> IO ())
message, data

-> Ptr a
data

-> IO ()

setErrorCallback

Arguments

:: Ptr Ctx

-> FunPtr (CString -> Ptr a -> IO ())
message, data

-> Ptr a
data

-> IO ()

ecPubKeyParse

Arguments

:: Ptr Ctx

-> Ptr PubKey64

-> Ptr CUChar
encoded public key array

-> CSize
size of encoded public key array

-> IO Ret

ecPubKeySerialize

Arguments

:: Ptr Ctx

-> Ptr CUChar
array for encoded public key, must be large enough

-> Ptr CSize
size of encoded public key, will be updated

-> Ptr PubKey64

-> SerFlags

-> IO Ret

ecdsaSignatureParseCompact :: Ptr Ctx -> Ptr Sig64 -> Ptr CompactSig -> IO Ret

ecdsaSignatureParseDer

Arguments

:: Ptr Ctx

-> Ptr Sig64

-> Ptr CUChar
encoded DER signature

-> CSize
size of encoded signature

-> IO Ret

ecdsaSignatureSerializeDer

Arguments

:: Ptr Ctx

-> Ptr CUChar
array for encoded signature, must be large enough

-> Ptr CSize
size of encoded signature, will be updated

-> Ptr Sig64

-> IO Ret

ecdsaSignatureSerializeCompact :: Ptr Ctx -> Ptr CompactSig -> Ptr Sig64 -> IO Ret

ecdsaVerify :: Ptr Ctx -> Ptr Sig64 -> Ptr Msg32 -> Ptr PubKey64 -> IO Ret

ecdsaSignatureNormalize

Arguments

:: Ptr Ctx

-> Ptr Sig64
output

-> Ptr Sig64
input

-> IO Ret

laxDerParse :: Ptr Ctx -> Ptr Sig64 -> Ptr CUChar -> CSize -> IO Ret

ecdsaSign

Arguments

:: Ptr Ctx

-> Ptr Sig64

-> Ptr Msg32

-> Ptr SecKey32

-> FunPtr (NonceFunction a)

-> Ptr a
nonce data

-> IO Ret

ecSecKeyVerify :: Ptr Ctx -> Ptr SecKey32 -> IO Ret

ecPubKeyCreate :: Ptr Ctx -> Ptr PubKey64 -> Ptr SecKey32 -> IO Ret

ecSecKeyTweakAdd :: Ptr Ctx -> Ptr SecKey32 -> Ptr Tweak32 -> IO Ret

ecPubKeyTweakAdd :: Ptr Ctx -> Ptr PubKey64 -> Ptr Tweak32 -> IO Ret

ecSecKeyTweakMul :: Ptr Ctx -> Ptr SecKey32 -> Ptr Tweak32 -> IO Ret

ecPubKeyTweakMul :: Ptr Ctx -> Ptr PubKey64 -> Ptr Tweak32 -> IO Ret

contextRandomize :: Ptr Ctx -> Ptr Seed32 -> IO Ret

ecPubKeyCombine

Arguments

:: Ptr Ctx

-> Ptr PubKey64
pointer to public key storage

-> Ptr (Ptr PubKey64)
pointer to array of public keys

-> CInt
number of public keys

-> IO Ret